remove portland from feeds for now

+++ builder/build.py
@@ -97,29 +97,30 @@ def runOTP(otpInputDir, otpOutputDir):
     print(colored(f'All done at {outputPath}', 'green'))
     # symlink OTP_OUTPUT_DIR/datestr/ to OTP_OUTPUT_DIR/default/
     currentLinkPath = os.path.join(otpOutputDir, DEFAULT_GRAPH_NAME)
-    if os.path.exists(currentLinkPath):
-      if os.path.islink(currentLinkPath):
-        os.unlink(currentLinkPath)
-        os.symlink(dateStr, currentLinkPath)
-        print(f'linking to {currentLinkPath}')
-        print(f'To check by hand run')
-        print(f'  java -Xmx2G -jar {CURRENT_OTP_JAR} --router default --graphs graphs --server')
-      else:
-        print(colored(f'{currentLinkPath} is not a symlink, not removing and not updating', 'yellow'))
-        print(f'To check by hand run')
-        print(f'  java -Xmx2G -jar {CURRENT_OTP_JAR} --router {dateStr} --graphs graphs --server')
+    if os.path.exists(currentLinkPath) and os.path.islink(currentLinkPath):
+      os.unlink(currentLinkPath)
+
+    if os.path.exists(currentLinkPath) and not os.path.islink(currentLinkPath):
+      print(colored(f'{currentLinkPath} is not a symlink, not removing and not updating', 'yellow'))
+      print(f'To check by hand run')
+      print(f'  java -Xmx2G -jar {CURRENT_OTP_JAR} --router {dateStr} --graphs graphs --server')
+    else:
+      os.symlink(dateStr, currentLinkPath)
+      print(f'linking to {currentLinkPath}')
+      print(f'To check by hand run')
+      print(f'  java -Xmx2G -jar {CURRENT_OTP_JAR} --router default --graphs graphs --server')
     
     print('check that it is working in nyc with')
     print('http://localhost:8080/otp/routers/default/plan?fromPlace=40.700853,-73.947738&toPlace=40.741524,%20-73.989330')
 
 def main():
-  global globalLogger
-  globalLogger = MyGlobalLogger('feed-import-%s.log' % datetime.datetime.now().strftime('%Y-%m-%d-%H-%M'))
-  checkForTransitFeed()
-  checkForPython2()
-  checkDirectories()
-  checkOTP()
-  downloadFeeds()
+  # global globalLogger
+  # globalLogger = MyGlobalLogger('feed-import-%s.log' % datetime.datetime.now().strftime('%Y-%m-%d-%H-%M'))
+  # checkForTransitFeed()
+  # checkForPython2()
+  # checkDirectories()
+  # checkOTP()
+  # downloadFeeds()
   runOTP(OTP_INPUT_DIR, OTP_OUTPUT_DIR)
 
 if __name__ == '__main__':
